pyproject.toml
[tool.poetry]
name = "ruscord-backend"
version = "0.1.0"
description = ""
authors = ["Timofey Tsepennikov <tta@biznpro.ru>"]
readme = "README.md"
packages = [{include = "*", from="src"}]

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.2"
djangorestframework = "^3.15.2"
djangorestframework-simplejwt = "^5.3.1"
django-cors-headers = "^4.5.0"
pillow = "^11.0.0"
channels = "^4.1.0"
channels-redis = "^4.2.0"
daphne = "^4.1.2"
python-dotenv = "^1.1.0"
psycopg2 = "^2.9.10"
livekit = "^1.0.1"
livekit-api = "^1.0.1"
drf-nested-routers = "^0.94.1"
drf-spectacular = "^0.28.0"
datetime = "^5.5"
twisted = {extras = ["http2", "tls"], version = "^24.11.0"}

[tool.poetry.group.dev.dependencies]
mypy = "^1.10.1"
django-stubs = "^5.0.2"
djangorestframework-stubs = "^3.15.0"

[tool.poetry.group.deploy.dependencies]
daphne = "^4.1.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
